pyproject.toml
[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core>=1.0.0"]

[tool.black]
color = true
line-length = 88
skip-string-normalization = true
target-version = ['py310']

[tool.isort]
combine_as_imports = true
include_trailing_comma = true
known_first_party = "pydantic"
line_length = 88
multi_line_output = 3
profile = "black"

[tool.mypy]
check_untyped_defs = true
disallow_any_decorated = false
disallow_any_explicit = false
disallow_any_expr = false
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
explicit_package_bases = true
ignore_missing_imports = true
namespace_packages = true
no_implicit_optional = true
no_implicit_reexport = true
python_version = "3.10"
show_error_codes = true
strict_concatenate = true
strict_equality = true
strict_optional = true
warn_no_return = true
warn_redundant_casts = true
warn_unreachable = true
warn_unused_configs = true
warn_unused_ignores = true

[tool.poetry]
authors = ["Meryll Dindin <merylldin@gmail.com>"]
description = "Dockerized Bastion for AWS ecosystems"
name = "telomere"
readme = "README.md"
version = "0.1.0"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pydantic = {extras = ["email"], version = "^1.10.2"}
pymysql = "^1.0.2"
fastapi = "^0.88.0"
bson = "^0.5.10"
pymongo = "^4.3.3"
pyjwt = "^2.6.0"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"
isort = "^5.10.1"
mypy = "^0.991"
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
ruff = "^0.0.178"
toml-sort = "^0.20.1"
uvicorn = "^0.20.0"

[tool.pytest.ini_options]
addopts = "--no-header -rA"
pythonpath = ["."]

[tool.ruff]
extend-select = ["Q"]
line-length = 88
target-version = "py310"
